edit userId type to long

+++ demo/src/main/java/com/codecool/demo/model/TaskAdder.java
@@ -3,16 +3,18 @@ package com.codecool.demo.model;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 
 public class TaskAdder {
-    private final int userId;
+    private final long userId;
     private final Task task;
 
     @JsonIgnoreProperties(ignoreUnknown = true)
-    public TaskAdder(int userId, Task task) {
+    public TaskAdder(
+            long userId,
+            Task task) {
         this.userId = userId;
         this.task = task;
     }
 
-    public int getUserId() {
+    public long getUserId() {
         return userId;
     }
